requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
bcrypt==4.2.0
black==24.8.0
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
cryptography==43.0.1
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
flake8==7.1.1
h11==0.14.0
identify==2.6.1
idna==3.10
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.8.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.31.0
virtualenv==20.26.6
